refactor(shared): extract shared utilities
