+++ .changeset/extract-shared-utils.md
@@ -1,0 +1,5 @@
+---
+'@lapidist/design-lint': patch
+---
+
+move shared utilities to dedicated package
