+++ packages/core/src/rules/token-border-radius.ts
@@ -6,7 +6,7 @@ import {
   extractVarName,
   closestToken,
 } from '../core/token-utils.js';
-import { isStyleValue } from '../utils/style.js';
+import { isStyleValue } from '@lapidist/design-lint-shared';
 
 interface BorderRadiusOptions {
   units?: string[];
