+++ packages/core/src/rules/token-duration.ts
@@ -6,7 +6,7 @@ import {
   extractVarName,
   closestToken,
 } from '../core/token-utils.js';
-import { isStyleValue } from '../utils/style.js';
+import { isStyleValue } from '@lapidist/design-lint-shared';
 
 export const durationRule: RuleModule = {
   name: 'design-token/duration',
