+++ packages/core/src/rules/token-spacing.ts
@@ -6,7 +6,7 @@ import {
   extractVarName,
   closestToken,
 } from '../core/token-utils.js';
-import { isStyleValue } from '../utils/style.js';
+import { isStyleValue } from '@lapidist/design-lint-shared';
 
 interface SpacingOptions {
   base?: number;
