+++ packages/shared/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@lapidist/design-lint-shared",
+  "version": "0.0.0",
+  "type": "module",
+  "exports": "./dist/index.js",
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "build": "tsc -p tsconfig.json",
+    "test": "tsx --test ../../tests/**/*.test.ts"
+  }
+}
