+++ packages/shared/src/index.ts
@@ -1,0 +1,2 @@
+export { isStyleValue } from './utils/style.js';
+export { isInNonStyleJsx } from './utils/jsx.js';
