+++ packages/shared/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "dist",
+    "typeRoots": ["./src/types", "../../node_modules/@types"],
+    "composite": true
+  },
+  "include": ["src"],
+  "exclude": ["dist"]
+}
