+++ tsconfig.eslint.json
@@ -4,5 +4,5 @@
     "rootDir": ".",
     "noEmit": true
   },
-  "include": ["src/**/*", "tests/**/*", "packages/core/src/**/*"]
+  "include": ["src/**/*", "tests/**/*", "packages/*/src/**/*"]
 }
